Close resources in a try stmt in reverse order
This also evaluates the finally block after the resources have been
closed.

This should be considered a workaround as it does not handle the
case where there is more than one resource and when exception is
thrown after the first resource is initialized but before the try body
is evaluated.